feat: Map boundaries. fix: Map upside down.
Before this commit the map was show with the 0,0 position on the top left corner, as the positions of the arrays that constitute the matrix. So when you went down, the Y axis increased. Now the 0,0 is in the lower left corner, as it is in the API.
Boundaries were also added to the map.